Improves naming by renaming ConfigCat.Client.Evaluate namespace to ConfigCat.Client.Evaluation (only refactoring without actual changes to logic or public API)
