+++ src/ConfigCat.Client.Tests/VariationIdEvaluatorTests.cs
@@ -1,4 +1,4 @@
-﻿using ConfigCat.Client.Evaluate;
+﻿using ConfigCat.Client.Evaluation;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using System;
