+++ src/ConfigCatClient/Evaluation/EvaluateResult.cs
@@ -4,7 +4,7 @@ using Newtonsoft.Json.Linq;
 using System.Text.Json;
 #endif
 
-namespace ConfigCat.Client.Evaluate
+namespace ConfigCat.Client.Evaluation
 {
     internal class EvaluateResult
     {
