+++ src/ConfigCatClient/Override/LocalDictionaryDataSource.cs
@@ -1,4 +1,4 @@
-﻿using ConfigCat.Client.Evaluate;
+﻿using ConfigCat.Client.Evaluation;
 using System;
 using System.Collections.Generic;
 using System.Linq;
